refactor(runner): fold BatchOperationsBar into the StatusStrip; demote wrapper-tools count
Two related Terminal-page chrome cleanups so the top strip owns "what needs
attention + how to respond" and nothing persistent floats over the grid.

1. Bulk actions move into the StatusStrip (BatchOperationsBar -> BatchActions).
   The old bar floated `absolute bottom-4` over the zone grid (occluding the
   bottom-center terminal) and showed its own "N waiting" count — duplicating
   the strip's "N need input" pill. The actions (Approve all / Reject all /
   Broadcast / Select, plus per-zone selection + the `{zone}/{n}/{title}/{tag}`
   broadcast templating) now render INLINE right after that pill via the new
   `BatchActions`, co-located with the count they act on and out of the grid.
   The broadcast input drops DOWN as an on-demand popover (the strip is
   top-pinned), so it overlays the grid only while open. The dismiss button is
   gone — the strip already auto-hides when nothing needs attention.

   Approve-all blind-sends `y` to every waiting session; now that it sits in the
   prominent top strip, a batch of >= 3 arms a one-shot inline confirm
   ("Approve N?") on the first click instead of firing immediately, so a
   reflexive click can't rubber-stamp many sessions unseen.

2. Wrapper-tools count demoted to an icon-only affordance. The raw "N wrapper
   tools" text was low-signal and rarely actionable, yet — because `wrapperCount`
   was in the `hasContent` gate and wrappers exist almost every session — it kept
   the strip permanently pinned open, defeating the auto-hide-when-idle design.
   The Package icon now opens the same popover (count lives on hover + in the
   popover header), and wrapper presence no longer keeps the strip open: it
   renders only when the strip is already up for a real signal. Pill skips the
   empty text span for the icon-only case.

Dead code removed with the float: the `BatchOperationsBar` mount + import in
TerminalPage, its `index.ts` export (now `BatchActions`), and the
`batchBarDismissed` state + its un-dismiss effect + `prevNeedsInputCountRef` in
useStateTransitionEffects.

tsc --noEmit clean; eslint clean for the changed files (the one Date.now purity
warning in StatusStrip's count memo is pre-existing and untouched).

Session-Id: 7c0aee67-7789-4ab6-8dee-4c7790f06bfe
Session-Name: runner-batchbar-into-statusstrip
